third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/websocket.https-expected.txt

CONSOLE ERROR: line 187: WebSocket connection to 'wss://web-platform.test:9444/echo' failed: Error during WebSocket handshake: Unexpected response code: 404
This is a testharness.js-based test.
FAIL Verify WebSocket handshake channel does not get intercepted assert_unreached: unexpected rejection: [object Event] Reached unreachable code
Harness: the test ran to completion.
